Fixed Null Liquids
Recipes without liquid outputs were causing errors. Fixed that by returning null instead of a liquid with a blank name and 0 quantity. I'm leaving the error that will occur if someone tries to have a quantity higher than 0 paired with a name that is incorrect, as that should give the user some feedback about the issue.